Note: This is a commit of work in progress...
Added ui components and code logic for user selectable canvas background colour. All infrastructure is in place but background is still rendering white for me thus this is a work in progress.

This commit also includes preliminary code to render vector polygons semi transparently but it causes quite a slowdown in drawing so Ive hard coded it to be opague until I have implement a gui frontend for setting transparency levels.



git-svn-id: http://svn.osgeo.org/qgis/trunk@4670 c8812cc2-4d05-0410-92ff-de0c093fc19c