Better fix for the issues raised in r5618 and r5626 - the selected item is now scrolled into view, even when the projection selector is being opened.
There still seems to be a slight issue when showing the Project Properties for the first time on a blank project (i.e. it scrolls *near* the WGS 84 entry, but not quite on top of it), but I'll leave that for another time.  Other situations now seem to work OK.



git-svn-id: http://svn.osgeo.org/qgis/trunk@5642 c8812cc2-4d05-0410-92ff-de0c093fc19c