Reworked the BandAligner tool to combine the estimate and apply steps into one.
This new function will load the whole source image into a memory buffer IF it is less than 256MiB. This resulted in an major speed improvement (740 seconds down to 473 seconds). Larger images will fall back to reading directly from disk.

Another improvement is the use of bicubic interpolation to create the disparity map. The original bi-linear interpolation resulted in visible "lines" in the disparity map which show up in the histogram of the disparity map.

There is still an issue with the findRegion function. The sub-pixel accuracy of this correlation function is not to great, also showing up in the histogram of the disparity map.

Updated the expected result of the band aligner test.
Also added a test that generate an image with an offset and verify that the output is close to those values.