Reworked the correlation code to give more accurate sub-pixel values.
The current implementation is arbitrarily fixed to a 0.1 sub-pixel accuracy. The resulting disparity map for our Sumbandila test image is now much more smoother, as one would expect.

Also improved the processing time by not writing the disparity map to disc if not required. Since the two disparity map files are 463MB in size for Sumbandila images, this can shave quite a few seconds of the processing time.

The data in the disparity grid file is now split into separate bands (no more complex float values). First band is the column shift (X), followed by the row shift (Y). After that comes the normalized RMS error of the algorithm. Finally the diff-phase value.